+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - API-9698: Refresh ES index after creating a product from the UI in order to well send product created event to event subscriptions
 - PIM-9711: Check that a category root isn't linked to a user or a channel before moving it to a sub-category
 - PIM-9730: Fix category tree initialization in the PEF when switching tabs
+- PIM-9679: Clean existing text attribute values removing linebreaks
 
 ## New features